flaskAPP/web/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
mysql-connector-python==8.3.0
gunicorn==22.0.0
